requirements.txt
hikari-lightbulb==2.2.4
python-dotenv==0.20.0
func_timeout==4.3.5
